feat: some small refactoring
