+++ components/ads-client/src/lib.rs
@@ -24,7 +24,7 @@ pub mod telemetry;
 
 pub use ffi::*;
 
-use crate::ffi::telemetry::MozAdsTelemetryWrapper;
+use crate::{ffi::telemetry::MozAdsTelemetryWrapper, worker::DispatchCommand};
 #[cfg(test)]
 mod test_utils;
 
@@ -175,45 +175,4 @@ impl MozAdsClient {
         self.command_tx.send(command).map_err(ComponentError::Dispatch)?;
         Ok(())
     }
-}
-
-#[derive(uniffi::Enum)]
-
-pub enum DispatchCommand {
-    RequestTileAd {
-        moz_ad_requests: Vec<MozAdsPlacementRequest>,
-        options: Option<MozAdsRequestOptions>,
-        callback: Arc<dyn TileRequestCallback>,
-    }
-}
-
-impl DispatchCommand {
-
-    #[handle_error(ComponentError)]
-    pub fn run_command(self, ads_client_inner : &MozAdsClientInner) -> AdsClientApiResult<()> {
-        // TODO: Duplicated behavior with the sync functions- either they call this or you call those
-        match self {
-            DispatchCommand::RequestTileAd { moz_ad_requests, options, callback } => {
-                let inner = ads_client_inner.lock();
-                let requests: Vec<AdPlacementRequest> = moz_ad_requests.iter().map(|r| r.into()).collect();
-                let options = options.unwrap_or_default();
-                let flags = AdRequestFlags::from(&options);
-                let ohttp = options.ohttp;
-                let cache_policy: CachePolicy = options.into();
-                let response = inner
-                    .request_tile_ads(requests, flags, Some(cache_policy), ohttp)
-                    .map_err(ComponentError::RequestAds)?;
-                let tiles = response.into_iter().map(|(k, v)| (k, v.into())).collect();
-                callback.on_ad(tiles);
-                // TODO: error, modular version
-            }
-        }
-        Ok(())
-    }
-}
-
-#[uniffi::export(with_foreign)]
-pub trait TileRequestCallback : Send + Sync {
-    fn on_ad(&self, tiles : HashMap<String, MozAdsTile>);
-    fn on_error(&self, err: MozAdsClientApiError);
 }
